+++ matching/pricing.go
@@ -1,15 +1,15 @@
 package matching
 
 const (
-	PRICE_LEVEL_DEFAULT = -1
-	PRICE_LEVEL_0 = 0
-	PRICE_LEVEL_1 = 10
-	PRICE_LEVEL_2 = 30
-	PRICE_LEVEL_3 = 50
-	PRICE_LEVEL_4 = 100
+	PRICE_LEVEL_DEFAULT = -1.0
+	PRICE_LEVEL_0 = 0.0
+	PRICE_LEVEL_1 = 10.0
+	PRICE_LEVEL_2 = 30.0
+	PRICE_LEVEL_3 = 50.0
+	PRICE_LEVEL_4 = 100.0
 )
 
-func checkPrice(priceLevel int) int{
+func checkPrice(priceLevel int) float64{
 	switch priceLevel {
 	case 0:
 		return PRICE_LEVEL_0
